Updated CVs
I am an IT Professional with over 15 years of experience in developing responsive, user-centric web applications. Proficient in modern web 
technologies and frameworks, with a passion for leveraging innovative solutions to solve complex problems. Known for strong collaboration, a positive 
attitude, and a commitment to excellence, I excel in creating scalable, maintainable applications, optimising code, and driving projects to successful 
completion.